movie_recommender-main/app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.4
    # via streamlit
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
cryptography==46.0.4
    # via google-auth
cuda-bindings==12.9.4
    # via torch
cuda-pathfinder==1.3.3
    # via cuda-bindings
distro==1.9.0
    # via google-genai
filelock==3.20.3
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2026.1.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-auth==2.48.0
    # via google-genai
google-genai==1.60.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
huggingface-hub==0.36.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.5.3
    # via scikit-learn
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
loguru==0.7.3
    # via -r requirements.in
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
narwhals==2.15.0
    # via altair
networkx==3.6.1
    # via torch
numpy==2.4.1
    # via
    #   -r requirements.in
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.4.5
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
packaging==26.0
    # via
    #   altair
    #   huggingface-hub
    #   streamlit
    #   transformers
pandas==2.3.3
    # via streamlit
pillow==12.1.0
    # via streamlit
protobuf==5.29.5
    # via streamlit
psycopg2-binary==2.9.11
    # via -r requirements.in
pyarrow==23.0.0
    # via streamlit
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   google-genai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.1.15
    # via transformers
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
    #   huggingface-hub
    #   streamlit
    #   transformers
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via sentence-transformers
scipy==1.17.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.2.0
    # via -r requirements.in
setuptools==80.10.1
    # via torch
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via google-genai
streamlit==1.53.1
    # via -r requirements.in
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via
    #   google-genai
    #   streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.10.0
    # via sentence-transformers
tornado==6.5.4
    # via streamlit
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.57.6
    # via sentence-transformers
triton==3.6.0
    # via torch
typing-extensions==4.15.0
    # via
    #   altair
    #   google-genai
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
watchdog==6.0.0
    # via streamlit
websockets==15.0.1
    # via google-genai
